+++ CHANGES.md
@@ -27,7 +27,7 @@ Changes in 12.8
 - Remove Database Backup program (secure by default) in DatabaseBackup.php, School_Setup/Menu.php, rosariosis.sql, rosariosis_mysql.sql
 - Remove $DatabaseDumpPath global variable in diagnostic.php, config.inc.sample.php, INSTALL.md, INSTALL_es.md & INSTALL_fr.md
 - Remove check for passthru function in Portal.php
-- Use SQLLimitForList() function in AddDrop.php
+- Use SQLLimitForList() function in AddDrop.php & ClassSearchWidget.fnc.php
 - Use DBUpdate() function, program wide
 
 Changes in 12.7.4